Add missing dep react-collapsible, clean up build.sh a bit

+++ build.sh
@@ -2,23 +2,27 @@
 trap restore_contract_state INT
 
 # make src/contracts a real directory instead of symlink
-rm src/contracts
-mkdir src/contracts
-cp build/contracts/* src/contracts
+if [ -h src/contracts ]; then #if it's a symlink
+    rm src/contracts
+    mkdir src/contracts #make it a directory
+fi
+if [ -d build/contracts ]; then
+    cp build/contracts/*.json src/contracts/
+fi
 
 # build
 node_modules/react-scripts/bin/react-scripts.js build || true
 
-function restore_contract_state() {
+restore_contract_state() {
     # move the built code into a production folder
     rm -rf production
     mv build production
 
     # make the src/contracts directory back into a symlink
     mkdir -p build/contracts
-    mv src/contracts/* build/contracts
+    mv src/contracts/* build/contracts/
     rm -rf src/contracts
-    ln -s "../build/contracts" "src/contracts"
+    ln -s "../build/contracts/" "src/contracts"
 }
 
 restore_contract_state
